rename _qt.resources to _qt.qt_resources
