+++ .gitignore
@@ -120,8 +120,8 @@ venv.bak/
 .vscode/
 
 # these will get autogenerated
-napari/resources/_qt_resources*.py
-napari/resources/res.qrc
+_qt_resources*.py
+res.qrc
 # ignore all generated themed svgs
 napari/resources/themes
