+++ napari/_qt/__init__.py
@@ -39,7 +39,7 @@ if StrictVersion(QtCore.__version__) < StrictVersion('5.12.3'):
 
 from .event_loop import gui_qt
 from .qt_main_window import Window
-from .resources import import_resources
+from .qt_resources import import_resources
 from .widgets.qt_range_slider import QHRangeSlider, QVRangeSlider
 
 path, loader = import_resources()
