+++ napari/_qt/qt_resources/build_icons.py
