+++ napari/_qt/qt_resources/styles/00_base.qss
