+++ napari/_qt/qt_resources/styles/01_buttons.qss
