+++ napari/_qt/qt_resources/styles/02_custom.qss
